Fixed more Findbugs issues. Now 40 remain!
Former-commit-id: 4c4498eb2584a4a82c258848e8c0207819061055
Former-commit-id: ba7904e31377593ff8fba46822b0d0ea0ddac58c

+++ server/src/main/java/org/icgc/dcc/validation/report/ValidationErrorReport.java
@@ -121,7 +121,7 @@ public class ValidationErrorReport implements Serializable {
       for(ColumnErrorReport column : this.columns) {
         List<Long> newLines = Lists.newLinkedList();
         for(Long oldLine : column.getLines()) {
-          newLines.add(byteToLine.get(oldLine).longValue());
+          newLines.add(byteToLine.get(oldLine));
         }
         column.setLines(newLines);
       }
